Even the first test fails. Apparently the metasocket does not close properly (I was using FastClose but it stopped working apparently) so I am implementing a proper close. It compiles but fails because I change in CloseAllSubflows the vector over which I am iterating. To fix. Also there is the problem that SendPendingData tries to resend acknowledge data. It shouldn't. Same subflow flags as NewAck an that was already acquitted :/

+++ src/internet/model/tcp-option-mptcp.cc
@@ -259,8 +259,8 @@ TcpOptionMpTcpCapable::SetRemoteKey(const uint64_t& remoteKey)
 void
 TcpOptionMpTcpCapable::Print (std::ostream &os) const
 {
-  os << "MP_CAPABLE. version" << m_version
-    << "MPTCP Flags: [" << m_flags << "]"
+  os << "MP_CAPABLE. version" << (int)m_version
+    << "MPTCP Flags: [" << (int)m_flags << "]"
     << "Sender's Key :[" << GetSenderKey() << "]"
     << "Peer's Key [";
   if( HasReceiverKey() )
